Reduced Automatron to a single header file, added namespaces
This allows its insertion into any test modules without the need for dependencies or installing Automatron as a plugin
