--- Source/Automatron/Private/Automatron.cpp
@@ -1,7 +1,0 @@
-// Copyright 2020 Splash Damage, Ltd. - All Rights Reserved.
-
-#include "Automatron.h"
-
-#if WITH_DEV_AUTOMATION_TESTS
-
-#endif //WITH_DEV_AUTOMATION_TESTS
