--- Source/Automatron/Private/Misc/Macros.h
@@ -1,11 +1,0 @@
-
-#pragma once
-
-#include <CoreMinimal.h>
-
-
-#define MakeSure(Condition) \
-    if(!ensure(Condition)) return
-
-#define MakeSureMsg(Condition, Format, ...) \
-    if(!ensureMsgf(Condition, Format, ##__VA_ARGS__)) return
