+++ Source/Automatron/Public/AutomatronModule.h
@@ -4,13 +4,10 @@
 
 #include <CoreMinimal.h>
 #include <Modules/ModuleManager.h>
-#include "TestSpec.h"
 
 
 class FAutomatronModule : public IModuleInterface
 {
-	static TArray<TSharedRef<FTestSpec>> SpecInstances;
-
 public:
 
 	/** Begin IModuleInterface implementation */
